requirements.txt
# KOSMOS Agent — Dependencies
# =============================
# Motor Cognitivo com Firecracker MicroVM Sandbox + DeepSeek LLM

# Core
numpy>=1.24.0

# Memória vetorial
faiss-cpu>=1.7.4

# Executor Jupyter
jupyter_client>=8.0.0
ipykernel>=6.0.0

# LLM (DeepSeek — OpenAI-compatible API)
requests>=2.28.0

# Testes (opcional)
# pytest>=7.0.0
